+++ Data/EditorProperties.cpp
@@ -2,6 +2,6 @@
 
 using namespace eLibV2::Data;
 
-unsigned int EditorProperties::mNumPages = 0;
-std::map<unsigned int, unsigned int> EditorProperties::mBackgroundBitmaps;
-std::map<std::string, unsigned int> EditorProperties::mControlBitmaps;
+UInt8 EditorProperties::mNumPages = 0;
+std::map<UInt8, UInt16> EditorProperties::mBackgroundBitmaps;
+std::map<std::string, UInt16> EditorProperties::mControlBitmaps;
