+++ Data/PluginProgram.cpp
@@ -4,19 +4,19 @@ using namespace eLibV2::Data;
 
 void PluginProgram::addParameter(const float value)
 {
-    if ((int)mParameterValues.size() < mNumParameters)
+    if ((UInt16)mParameterValues.size() < mNumParameters)
         mParameterValues.push_back(value);
 }
 
-void PluginProgram::setParameter(const unsigned long parameterIndex, const float value)
+void PluginProgram::setParameter(const UInt16 parameterIndex, const float value)
 {
-    if ((unsigned long)mParameterValues.size() > parameterIndex)
+    if ((UInt16)mParameterValues.size() > parameterIndex)
         mParameterValues[parameterIndex] = value;
 }
 
-float PluginProgram::getParameter(const unsigned long parameterIndex) const
+float PluginProgram::getParameter(const UInt16 parameterIndex) const
 {
-    if ((unsigned long)mParameterValues.size() > parameterIndex)
+    if ((UInt16)mParameterValues.size() > parameterIndex)
         return mParameterValues[parameterIndex];
     else
         return 0.0f;
